Refactor: Define interfaces for Storer and Manager
