+++ internal/storage/postgres/db.go
@@ -5,17 +5,8 @@
 package postgres
 
 import (
-	"context"
-
 	"github.com/jackc/pgx/v5"
-	"github.com/jackc/pgx/v5/pgconn"
 )
-
-type DBTX interface {
-	Exec(context.Context, string, ...interface{}) (pgconn.CommandTag, error)
-	Query(context.Context, string, ...interface{}) (pgx.Rows, error)
-	QueryRow(context.Context, string, ...interface{}) pgx.Row
-}
 
 func New(db DBTX) *Queries {
 	return &Queries{db: db}
